requirements-dev.txt
paho-mqtt==1.6.1
pytest==7.4.0
coverage==7.3.0
